storage capacity: move node selection to status
There may be reasons for updating the set of nodes without replacing
the entire CSIStoragePool, therefore it's better to have this information
in the status. One effect is that different storage pools now will have
an identical spec because it's just the driver name, so the API no longer
requires that the spec is unique. The external-provisioner still ensures
that it doesn't create more objects than necessary.

While at it, some formatting (spaces instead of tabs) and a
cut-and-paste error get fixed.