dynamic resource allocation: introduce PodScheduling object
This replaces the individual ResourceClaimScheduling objects.  The advantage is
that the PotentialNodes list only needs to be set once. The scheduler logic is
still the same (try allocation immediately, retry based on new information
later).

The new object also has the advantage that it provides a patch towards more
cooperative scheduling with different drivers reacting to the information
published by others, should that become necessary or desirable.