KEP 4381: semantic model, resource attributes, in-tree parameter type
The terms "numeric model", "node resource capacity", "capacity type" were
emphasizing the numeric aspect of resources too much. A better mental model is
to think of a "semantic model" as describing resources with "attributes" which
can be both qualitative and quantitative. Collecting that information for from
all drivers on a node leads to the "node resources". This is not a conceptual
change, just different language to explain the original idea better.

The usage of RawExtension gets replaced with strongly-typed fields in the
resource.k8s.io API. This adds better validation, but has the downside that now
kubelet is tightly coupled with the control plane. The kubelet must support all
versions of semantic models that a DRA driver might want to use. Building a
custom scheduler with its own semantic model is no longer supported.

Converting from a vendor CRD into the in-tree representation of claim
parameters inside the scheduler was problematic, no matter how this would have
been done. The approach take now is to require that the vendor writes a CRD
controller which converts from the vendor CRD into a ResourceClaimParameters
object which contains the in-tree claim parameters representation as defined by
the semantic models that the vendor is using.