KEP 4381: switch to "partitionable" model, class parameters, avoid reading CRD
The big change is the switch to a single semantic model that can handle GPUs by
listing instances and valid partitioning of those.

Instead of requiring that the scheduler reads CRDs and copies those to the
allocation result, it is now the DRA driver which copies the additional
configuration parameters in some vendor type into the generated parameters. By
doing the same for class parameters, it becomes possible to support additional
filtering via class parameters. This was excluded before.